rename promise-queue to lastQueue
