+++ client/src/components/providers/SingleQueryProvider.js
@@ -1,5 +1,5 @@
 import hash from "object-hash";
-import { LastQueue } from "utils/promise-queue";
+import { LastQueue } from "utils/lastQueue";
 
 import { HasAttributesMixin } from "./utils";
