+++ client/src/stores/historyItemsStore.ts
@@ -11,7 +11,7 @@ import Vue, { computed, ref } from "vue";
 import type { DatasetSummary, HDCASummary } from "@/api";
 import { HistoryFilters } from "@/components/History/HistoryFilters";
 import { mergeArray } from "@/store/historyStore/model/utilities";
-import { LastQueue } from "@/utils/promise-queue";
+import { LastQueue } from "@/utils/lastQueue";
 import { urlData } from "@/utils/url";
 
 type HistoryItem = DatasetSummary | HDCASummary;
